// src/generated/models/Cr9cd_teamsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr9cd_teamscr9cd_default_platform = {
  100000000: 'Mock',
  100000001: 'Ticketmaster',
  100000002: 'AXS',
  100000003: 'SeatGeek'
} as const;
export type Cr9cd_teamscr9cd_default_platform = keyof typeof Cr9cd_teamscr9cd_default_platform;
export const Cr9cd_teamsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr9cd_teamsstatecode = keyof typeof Cr9cd_teamsstatecode;
export const Cr9cd_teamsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr9cd_teamsstatuscode = keyof typeof Cr9cd_teamsstatuscode;

export interface Cr9cd_teamsBase {
  cr9cd_abbreviation?: string;
  cr9cd_default_platform?: Cr9cd_teamscr9cd_default_platform;
  cr9cd_default_tickets_per_game?: number;
  cr9cd_home_games_per_season?: number;
  cr9cd_name?: string;
  cr9cd_official_url?: string;
  cr9cd_sport?: string;
  cr9cd_teamid: string;
  cr9cd_venue?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr9cd_teamsstatecode;
  statuscode?: Cr9cd_teamsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr9cd_teams extends Cr9cd_teamsBase {
  cr9cd_default_platformname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}