// src/generated/models/Cr9cd_seasonsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr9cd_seasonscr9cd_status = {
  100000000: 'Draft',
  100000001: 'Active',
  100000002: 'Archived'
} as const;
export type Cr9cd_seasonscr9cd_status = keyof typeof Cr9cd_seasonscr9cd_status;
export const Cr9cd_seasonsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr9cd_seasonsstatecode = keyof typeof Cr9cd_seasonsstatecode;
export const Cr9cd_seasonsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr9cd_seasonsstatuscode = keyof typeof Cr9cd_seasonsstatuscode;

export interface Cr9cd_seasonsBase {
  cr9cd_end_date?: string;
  cr9cd_name?: string;
  cr9cd_seasonid: string;
  cr9cd_start_date?: string;
  cr9cd_status?: Cr9cd_seasonscr9cd_status;
  "cr9cd_Team@odata.bind": string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr9cd_seasonsstatecode;
  statuscode?: Cr9cd_seasonsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr9cd_seasons extends Cr9cd_seasonsBase {
  cr9cd_statusname?: string;
  cr9cd_teamname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr9cd_team?: object;
  _cr9cd_team_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}